+++ contracts/transfer/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added support for Economic Protocol scenario 3 [#1630]
+- Added method which exposes the current gas price [#1604]
+
 ### Changed
 
 - Change dependencies declarations enforce bytecheck [#1371]
@@ -17,5 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.7.0] - 2023-12-15
 
+[#1630]: https://github.com/dusk-network/rusk/issues/1630
+[#1604]: https://github.com/dusk-network/rusk/issues/1604
 [#1675]: https://github.com/dusk-network/rusk/issues/1675
 [#1371]: https://github.com/dusk-network/rusk/issues/1371